requirements.txt
tensorflow
numpy
matplotlib
opencv-python
pillow
